[REFACT]: 수정사항 반영 (PresentationModel 분리 및 파라미터로 값 전달)
