+++ Projects/Presentation/Challenge/Implementations/PresentationModel/ChallengeModifyPresentaionModel.swift
@@ -1,0 +1,17 @@
+//
+//  ChallengeModifyPresentaionModel.swift
+//  Presentation
+//
+//  Created by 임우섭 on 6/3/25.
+//  Copyright © 2025 com.photi. All rights reserved.
+//
+
+struct ModifyPresentationModel {
+  let title: String
+  let hashtags: [String]
+  let verificationTime: String
+  let goal: String
+  let imageUrlString: String
+  let rules: [String]
+  let deadLine: String
+}
