+++ Projects/Presentation/Challenge/Implementations/PresentationModel/ChallengePreviewPresentationModel.swift
@@ -1,0 +1,19 @@
+//
+//  ChallengePreviewPresentationModel.swift
+//  Presentation
+//
+//  Created by 임우섭 on 6/3/25.
+//  Copyright © 2025 com.photi. All rights reserved.
+//
+
+import Core
+
+struct PreviewPresentationModel {
+  let title: String
+  let hashtags: [String]
+  let verificationTime: String
+  let goal: String
+  let image: UIImageWrapper
+  let rules: [String]
+  let deadLine: String
+}
